requirements.txt
Pillow==8.1.0
requests==2.25.1
selenium==3.141.0
